backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
libsql-experimental==0.0.21
passlib==1.7.4
python-jose[cryptography]==3.3.0
google-generativeai==0.3.2
python-multipart==0.0.6
PyJWT==2.8.0
sqlalchemy